Move server code out of cli
