+++ packages/server/package.json
@@ -25,7 +25,7 @@
     "url": "https://twitter.com/flybayer"
   },
   "module": "dist/webpack.esm.js",
-  "typings": "dist/packages/webpack/src",
+  "typings": "dist/packages/server/src",
   "devDependencies": {
     "@types/jest": "^25.1.3",
     "husky": "^4.2.3",
