+++ packages/server/src/index.ts
@@ -1,1 +1,4 @@
 export {withBlitz} from './withBlitz'
+export {build} from './build'
+export {dev} from './dev'
+export {prod} from './prod'
