+++ packages/server/src/next-utils.ts
@@ -1,14 +1,14 @@
 import {spawn} from 'child_process'
 
 export async function nextStartDev(nextBin: string, cwd: string) {
-  return Promise.resolve(
-    spawn(nextBin, ['dev'], {
-      cwd,
-      stdio: 'inherit',
-    }).on('error', err => {
-      console.error(err)
-    }),
-  )
+  const cp = spawn(nextBin, ['dev'], {
+    cwd,
+    stdio: 'inherit',
+  }).on('error', err => {
+    console.error(err)
+  })
+
+  return Promise.resolve(cp)
 }
 
 export async function nextBuild(nextBin: string, cwd: string) {
